packages/execution/pyproject.toml
[project]
name = "graphweave-execution"
version = "0.1.0"
description = "GraphWeave execution layer — FastAPI + LangGraph"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langgraph>=1.0.0,<2.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "slowapi>=0.1.9",
    "httpx>=0.28.0",
    "simpleeval>=1.0.3",
    "trafilatura>=2.0.0",
    "pydantic>=2.11.0",
    "aiosqlite>=0.22.1",
    "tavily-python>=0.5.0,<1.0",
    "duckduckgo-search>=6.0.0,<8.0",
    "Wikipedia-API>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "langchain-core>=0.3.0",
    "ruff>=0.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in defaults is standard FastAPI pattern

[tool.ruff.format]
quote-style = "double"
